+++ JAADI2CLib.h
@@ -124,21 +124,21 @@ typedef enum {
 //Structs for data transfer
 
 typedef struct {
-	int x;
-    int y;
-    int z;
+	float x;
+    float y;
+    float z;
 } AccelData;
 
 typedef struct {
-	int x;
-    int y;
-    int z;
+	float x;
+    float y;
+    float z;
 } MagData;
 
 typedef struct {
-	int16_t x;
-    int16_t y;
-    int16_t z;
+	float x;
+    float y;
+    float z;
 } GyroData;
